requirements.txt
streamlit==1.29.0
PyPDF2==3.0.1
faiss-cpu==1.7.4

sentence-transformers==2.2.2
huggingface-hub==0.14.1
transformers==4.30.2

torch==2.1.1
pandas==2.1.3
numpy==1.24.3
pdfplumber==0.11.0